example server: support serving over stdio
This chnage adds an "stdio" flag to the example server in order to
support `remote` p11-kit module entries, which dynamically launch a
managed server for clients that support it. This includes clients that
use gnutls.

See pkcs11.conf(5) for details on what the remote entry is meant to do.