pkcs11: fix return value for invalid attributes
C_GetAttributeValue should return CKR_ATTRIBUTE_TYPE_INVALID in the case
that one or more of the attributes requested aren't valid for the given
object, so this fixes that.

When using gnutls this actually manifests as a segfault, as it assumes
in the CKR_OK case that all the returned buffers are valid, but p11-kit
"helpfully" sets the "length" value for invalid buffers to -1, which
gnutls interprets as a uint, which means it tries to read a very large
amount of memory before eventually failing.

Addresses part of #21.